+++ examples/resources/iosxe_cli/resource.tf
@@ -1,0 +1,6 @@
+resource "iosxe_cli" "example" {
+  cli = <<-EOT
+  interface Loopback123
+  description configured-via-restconf-cli
+  EOT
+}