+++ gen/templates/provider.go
@@ -278,6 +278,7 @@ func (p *IosxeProvider) Resources(ctx context.Context) []func() resource.Resourc
 	return []func() resource.Resource{
 		NewRestconfResource,
 		NewSaveConfigResource,
+		NewCliResource,
 		{{- range .}}
 		New{{camelCase .Name}}Resource,
 		{{- end}}