+++ internal/provider/provider.go
@@ -276,6 +276,7 @@ func (p *IosxeProvider) Resources(ctx context.Context) []func() resource.Resourc
 	return []func() resource.Resource{
 		NewRestconfResource,
 		NewSaveConfigResource,
+		NewCliResource,
 		NewAAAResource,
 		NewAAAAccountingResource,
 		NewAAAAuthenticationResource,