+++ docs/guides/changelog.md
@@ -12,6 +12,7 @@ description: |-
 - Add `iosxe_vlan_filter` resource and data source
 - Add `iosxe_vlan_group` resource and data source
 - Add `iosxe_save_config` resource
+- Add `iosxe_cli` resource
 
 ## 0.5.0